Fix defects in MySQL timeout and custom DNS resolver
* Also, improve error checking.

+++ types.go
@@ -21,7 +21,7 @@ type Site struct {
 	Protocol       string      `json:"protocol"`
 	HTTPConfig     HTTPConfig  `json:"http"`
 	MySQLConfig    MySQLConfig `json:"mysql"`
-	TimeoutSeconds int         `json:"timeoutSeconds"`
+	TimeoutSeconds int64       `json:"timeoutSeconds"`
 	Recipients     []string    `json:"recipients"`
 }
